requirements.txt
django-crispy-forms==1.9.2
django==3.1
python-decouple==3.3